% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_chat.R
\name{cli_chat}
\alias{cli_chat}
\title{CLI application utilizing the OpenAI chat endpoint}
\usage{
cli_chat(api_key)
}
\arguments{
\item{api_key}{string, OpenAI API key
(see https://platform.openai.com/account/api-keys)}
}
\description{
This CLI application allows to engage in text-based conversations
by interacting with ChatGPT.
You can send messages or questions to the chatbot via the command line
interface (CLI) and receive responses generated by the AI model.
Additionally, you have the flexibility to adjust various parameters
that govern the behavior of the chatbot.
}
\examples{
\dontrun{
  cli_chat("super-secret-key")
}

}
