% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_completions.R
\name{cli_completions_data_transmission}
\alias{cli_completions_data_transmission}
\title{CLI application utilizing the OpenAI completions endpoint with data transmission}
\usage{
cli_completions_data_transmission(api_key)
}
\arguments{
\item{api_key}{string, OpenAI API key
(see https://platform.openai.com/account/api-keys)}
}
\description{
CLI application utilizing the OpenAI completions endpoint with data transmission
}
\examples{
\dontrun{
  cli_completions_data_transmission("super-secret-key")
}

}
