% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_completions.R
\name{cli_completions_simle_data}
\alias{cli_completions_simle_data}
\title{CLI application utilizing the OpenAI completions endpoint with simple data}
\usage{
cli_completions_simle_data(api_key)
}
\arguments{
\item{api_key}{string, OpenAI API key
(see https://platform.openai.com/account/api-keys)}
}
\description{
This function sends four simple queries (simultaneously)
and then displays the answers provided by AI.
}
\examples{
\dontrun{
  cli_completions_simle_data("super-secret-key")
}

}
