% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp_modChat.R
\name{shinyapp_modChat}
\alias{shinyapp_modChat}
\title{Module chat (server)}
\usage{
shinyapp_modChat(id, api_key)
}
\arguments{
\item{id}{module id}

\item{api_key}{OpenAI api key}
}
\description{
Module chat (server)
}
