% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp_modFiles.R
\name{shinyapp_modFiles}
\alias{shinyapp_modFiles}
\title{Module files (server)}
\usage{
shinyapp_modFiles(id, api_key, files_df)
}
\arguments{
\item{id}{module id}

\item{api_key}{OpenAI api key}

\item{files_df}{files data frame (reactive value)}
}
\description{
Module files (server)
}
