% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp_modFineTuning.R
\name{shinyapp_modFineTuningUI}
\alias{shinyapp_modFineTuningUI}
\title{Module fine-tunes (UI)}
\usage{
shinyapp_modFineTuningUI(id)
}
\arguments{
\item{id}{module id}
}
\description{
Module fine-tunes (UI)
}
