% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp_modImageEdit.R
\name{shinyapp_modImageEdit}
\alias{shinyapp_modImageEdit}
\title{Module image edit (server)}
\usage{
shinyapp_modImageEdit(id, api_key)
}
\arguments{
\item{id}{module id}

\item{api_key}{OpenAI api key}
}
\description{
Module image edit (server)
}
