% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp_modImageEdit.R
\name{shinyapp_modImageEditUI}
\alias{shinyapp_modImageEditUI}
\title{Module image edit (UI)}
\usage{
shinyapp_modImageEditUI(id)
}
\arguments{
\item{id}{module id}
}
\description{
Module image edit (UI)
}
