% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp_modImageGenerator.R
\name{shinyapp_modImagesGenerator}
\alias{shinyapp_modImagesGenerator}
\title{Module images generator (server)}
\usage{
shinyapp_modImagesGenerator(id, api_key)
}
\arguments{
\item{id}{module id}

\item{api_key}{OpenAI api key}
}
\description{
Module images generator (server)
}
