% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp_modImageGenerator.R
\name{shinyapp_modImagesGeneratorUI}
\alias{shinyapp_modImagesGeneratorUI}
\title{Module images generator (UI)}
\usage{
shinyapp_modImagesGeneratorUI(id)
}
\arguments{
\item{id}{module id}
}
\description{
Module images generator (UI)
}
