% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp_utils.R
\name{shinyapp_modX_roxygen}
\alias{shinyapp_modX_roxygen}
\title{shinyapp_modX roxygen helper}
\usage{
shinyapp_modX_roxygen(id, api_key)
}
\arguments{
\item{id}{module id}

\item{api_key}{OpenAI api key}
}
\description{
shinyapp_modX roxygen helper
}
